Used PHP Behat configuration as behat.php when the edition ships one (#145)
* Used the PHP Behat configuration as behat.php when the edition ships one

* Fixed missing semicolon before then in the valkey branch of the 6.0 prepare script

* [TEMP] Listed Behat configuration files after copying

* Removed recipe-provided Behat configuration files before creating behat.php

* Fixed missing semicolon before then in the valkey branch of the stable prepare script

* Relied on the ibexabehat default configuration instead of removing recipe files

(cherry picked from commit f00b423d2e2c0c179817da684e1bd8b61337259f)

+++ bin/6.0.x-dev/prepare_project_edition.sh
@@ -161,7 +161,11 @@ docker exec install_dependencies composer update --no-scripts
 sudo sed -i "s/\['test' => true\]/\['test' => true, 'behat' => true\]/g" config/bundles.php
 
 # Create a default Behat configuration file
-cp "behat_ibexa_${PROJECT_EDITION}.yaml" behat.yaml
+if [ -f "behat_ibexa_${PROJECT_EDITION}.php" ]; then
+    cp "behat_ibexa_${PROJECT_EDITION}.php" behat.php
+else
+    cp "behat_ibexa_${PROJECT_EDITION}.yaml" behat.yaml
+fi
 
 # Depenencies are installed and container can be removed
 docker container stop install_dependencies
@@ -203,7 +207,7 @@ fi
 if [[ "$COMPOSE_FILE" == *"redis"*.yml ]]; then
     echo '> Display SPI (Redis) version for debugging'
     docker exec ibexa-redis-1 sh -c "redis-cli --version"
-elif [[ "$COMPOSE_FILE" == *"valkey"*.yml ]] then
+elif [[ "$COMPOSE_FILE" == *"valkey"*.yml ]]; then
     echo '> Display SPI (Valkey) version for debugging'
     docker exec valkey sh -c "valkey-cli --version"
 fi
